Reworked MediumDec
Finally compliable with operate code changes
-Still need to add back in the math features before testing and adjusting MediumDecV2 and AltDec (and later MixedDec)

+++ AltNum/TinyUDec.hpp
@@ -3306,9 +3306,9 @@ public:
         }
 
         TinyUDec PowConstOp(TinyUDec& targetValue, const int& expValue) { return IntPowConstOp(targetValue, expValue); }
-        TinyUDec PowConstOp(TinyUDec& targetValue, const long long& expValue) { return IntPowConstOp(targetValue, expValue); }
+        TinyUDec PowConstOp(TinyUDec& targetValue, const __int64& expValue) { return IntPowConstOp(targetValue, expValue); }
         TinyUDec PowConst(TinyUDec& targetValue, const int& expValue) { return IntPowConstOp(targetValue, expValue); }
-        TinyUDec PowConst(TinyUDec& targetValue, const long long& expValue) { return IntPowConstOp(targetValue, expValue); }
+        TinyUDec PowConst(TinyUDec& targetValue, const __int64& expValue) { return IntPowConstOp(targetValue, expValue); }
 
 
         /// <summary>
